LayoutTests/imported/w3c/web-platform-tests/scroll-animations/view-timelines/view-timeline-get-set-range-expected.txt


FAIL Getting and setting the animation range assert_equals: rangeStart set to scroll 10% expected (string) "scroll" but got (undefined) undefined
